requirements.txt
streamlit~=1.41.0
pillow~=11.0.0
torchvision~=0.20.1
torch~=2.5.1
opencv-python~=4.10.0.84
numpy~=2.2.0